vertex_training/requirements.txt
torch>=2.0
google-cloud-bigquery
google-cloud-storage
pandas
scikit-learn
db-dtypes
